// src/modules/channels/types/channel-error-code.enum.ts
export enum ChannelErrorCode {
  CHANNEL_NOT_FOUND = 'CHANNEL_NOT_FOUND',
  NOT_A_CHANNEL = 'NOT_A_CHANNEL',
  CHANNEL_PRIVATE_OR_NO_USERNAME = 'CHANNEL_PRIVATE_OR_NO_USERNAME',
  BOT_FORBIDDEN = 'BOT_FORBIDDEN',
  USER_NOT_ADMIN = 'USER_NOT_ADMIN',
  BOT_NOT_ADMIN = 'BOT_NOT_ADMIN',
  BOT_MISSING_RIGHTS = 'BOT_MISSING_RIGHTS',
  INVALID_USERNAME = 'INVALID_USERNAME',
  USER_NOT_MEMBER = 'USER_NOT_MEMBER',
  MEMBERSHIP_DISABLED = 'MEMBERSHIP_DISABLED',
  MEMBERSHIP_INACTIVE = 'MEMBERSHIP_INACTIVE',
  NOT_ADMIN = 'NOT_ADMIN',
  NOT_ADMIN_ANYMORE = 'NOT_ADMIN_ANYMORE',
  MISSING_RIGHTS = 'MISSING_RIGHTS',
  USER_NOT_CREATOR = 'USER_NOT_CREATOR',
  CHANNEL_ALREADY_LINKED = 'CHANNEL_ALREADY_LINKED',
  MODERATOR_NOT_ADMIN = 'MODERATOR_NOT_ADMIN',
}